updateCard in useHandleQA hook + updated Dashboard + similarQ logic moved to useHandleQA from useAddSingleQA
